Day10/advent-of-agents-spring-26/pyproject.toml
[project]
name = "aoa-luissala"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.26.0",
]
